pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pci"
version = "0.1.0"
description = "Personal Content Index"
requires-python = ">=3.11"
dependencies = [
    "typer",
    "rich",
    "sqlite-vec",
    "anthropic",
    "sentence-transformers",
    "trafilatura",
    "yt-dlp[default]",
    "webvtt-py",
    "sqlean.py",
    "python-dotenv>=1.2.2",
]

[project.scripts]
pci = "pci.cli:app"
